pyproject.toml
[project]
name = "redditide"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "httpx>=0.27.0",
    "pytest>=8.1.1",
    "pyyaml>=6.0.1",
    "pytest-httpx>=0.30.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = []
